Sync total tick count to client instead.
For sake of avoid unnecessary packet, the
criterion of triggering data sync in Container
("become dirty") is different data from last
known value. If we only sync number of minutes,
players will see "0 hour remain" if they open
the GUI shortly after close it, for the data aren't
updated in time.
As such, to ensure player always see the latest
data, we have to sync the time in ticks.